\name{convert.factors.to.strings.in.dataframe}
\alias{convert.factors.to.strings.in.dataframe}
\title{Convert data.frame columns from factors to characters}
\description{Convert data.frame columns from factors to characters
}
\usage{  
convert.factors.to.strings.in.dataframe(dataframe)
}  
   

\arguments{  
\item{dataframe}{ dataframe with columns of factors}
} 
 
       
\value{
\item{dataframe}{ dataframe with columns of characters}
}

  
 
 